work on float stuff mostly done.
results objects still seem to be returning float64s which indicates SFH.fit_sfh may be doing type promotion. But in any case, it is actually slower because the BFGs convergence is worse, so we should actually just always do float64 I think, unless we allow lower precision in the optimization as well. Several fits fail to converge after 1000 iterations, and iterating is slow -- priority should be to reduce the number of iterations.